pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ancient-pdf-master"
version = "1.0.0"
description = "OCR tool for Ancient Greek, Latin, and English texts with searchable PDF output"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "pdf2image>=1.16",
    "pikepdf>=8.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
